requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.23
scikit-learn>=1.2
xgboost>=1.7
pyyaml>=6.0
joblib>=1.2
